[Fix] 리뷰 피드백 반영
- COUNT 반환 타입(Long)에 맞추어 수정

+++ src/main/java/com/oneovern/domain/notification/controller/NotificationController.java
@@ -44,7 +44,7 @@ public class NotificationController {
 
     //읽지 않은 알림 개수 조회
     @GetMapping("/unread-count")
-    public ApiResponse<Integer> getUnreadNotificationCount(
+    public ApiResponse<Long> getUnreadNotificationCount(
             @AuthUser Member member
     )
     {
